Introduce a PackageBoundaryMode enum to FilesetTraversalParams.
This enables the introduction of a third behavior mode when
RecursiveFilesystemTraversalFunction encounters a package boundary: on top of
the existing cross/no-cross behavior, it can now bail out with an error.

This is preparatory work to support Skyframe-native Filesets.

The CL also contains a couple of prettifying cleanups (in ActionCacheChecker
and in FilesetManifestAction).

--
MOS_MIGRATED_REVID=86185105