remove ActionMetadata#getInputCount
Removing ActionMetadata#getInputCount instead of relying on all
implementations of ActionMetadata keeping getInputCount and
getInputs in sync. getInputCount is only used by a few tests
and has no applications in the main code at the moment, so it
is trivial to remove before we may come to rely on it more or its
semantics get more complicated.

--
MOS_MIGRATED_REVID=89130009