Cleanup: Build file.cc as part of util library.
Looks like in the transition it was forgotten to be moved to the new
location.

As an effect of this change it was discovered that file.cc had an
dependency back on blaze_util.cc (from client binary for die/pdie functions).

In order to fix that dependency we have had to move die/pdie functions into
util library (added errors.(cc|h)) for this.

Tested on Linux with the following command lines:

$ ./compile.sh
$ ./output/bazel build //src/main/cpp:all
$ ./output/bazel build //src:bazel
$ ./bootstrap_test.sh all

--
Change-Id: I70fb1b6189dc5af31f816332d76efe380ef26302
MOS_MIGRATED_REVID=90987514