Fix SkyframeExecutor to not assume only one node for the configuration(s).
As far as I can tell, the dependency on the BuildConfigurationCollection node
is not required anymore. The BuildConfiguration is part of the SkyKey, and it
uses object identity.

This shouldn't affect garbage collection due to the explicit
BuildView.dropConfiguredTargets. If we change that in the future, we will have
to be (even more) careful.

--
MOS_MIGRATED_REVID=92134490