Rollback of commit 77e28ae310d5e6731248340e5e037168d2de1a41.
*** Reason for rollback ***

The double quotes are actually incorrect, because the strings we pass in are already single quoted.

*** Original change description ***

Update ios_runner template to dynamically receive options

The SDK_VERSION and SIM_DEVICE variables in the runner script had
hard-coded values. This change updates them to be set by the
template expansion, so they can be controlled by the command-line
options.

--
MOS_MIGRATED_REVID=92337057