Fix race condition in MemoizingEvaluatorTest#cycleAndErrorAndReady.
If otherTop did not enqueue its dep for evaluation before errorKey threw, its dep would never be enqueued and therefore never signal. This used to be ok because we didn't shut the threadpool down as aggressively.

The fix is just to delay the actual throwing of errorKey until otherTop is signaled.

--
MOS_MIGRATED_REVID=93349168