Add a converter for enum flags that can also be boolean.
A new converter class is added that makes it possible for enum flags to define a conversion from booleans to one of the enumeration values. This is in addition to the enumeration conversions.

Fields that define such a converter can be used with the --flag and --noflag
forms in addition to their --flag=value enumeration style.

--
MOS_MIGRATED_REVID=93972718