Allow objc_{library,binary} to depend on cc_library.
This is an early version of support for this feature, likely still missing
a number of edge cases. However the basic functionality should work.

To allow a dependency from objc to cc, the following flags will have to be
passed to bazel:

--experimental_enable_objc_cc_deps --experimental_disable_java --cpu=ios_i386 --crosstool_top=//tools/objc/crosstool:crosstool

The feature is also compatible with --ios_multi_cpus, with the familiar
values for --ios_cpu (i386, x86_64, armv7, arm64). However, using this
crosstool and any CPU defined in it (legal values are ios_i386, ios_x86_64,
ios_armv7, ios_arm64) will make it impossible to use genrules with java make 
variables in the same build: Obviously they require java support and no that
is not available for iOS CPUs.

The new flag --experimental_disable_java has not been tested in any
circumstances but the one enabled by this CL so use it with caution.

This CL does not contain any Xcode support for CC dependencies yet, they
will just not show up in the generated Xcode project.

--
MOS_MIGRATED_REVID=94116942