Rollback of commit 6217be8606065f126c42b03350292a74798d1281.
*** Reason for rollback ***

This was accidentally submitted (why did tappresubmit not complain about no LGTM?!)

*** Original change description ***

Allow objc_{library,binary} to depend on cc_library.

This is an early version of support for this feature, likely still missing
a number of edge cases. However the basic functionality should work.

To allow a dependency from objc to cc, the following flags will have to be
passed to bazel:

--experimental_enable_objc_cc_deps --experimental_disable_java --cpu=ios_i386 --crosstool_top=//tools/objc/crosstool:crosstool

The feature is also compatible with --ios_multi_cpus, with the familiar
values f...

***

--
MOS_MIGRATED_REVID=94118959