Add --with_aspect_deps flag to blaze query. This flag should produce additional information about aspect dependencies when --output is set to {xml, proto}.
One quirk of this CL is that if BUILD files of direct dependencies are added both under <subinclude> and <load>. Any better ideas are appreciated. 

As a drive-by fix, if for some reason a package reports the same label as a both subinclude and a Skylark dependency, it will only be reported once in the proto output.

RELNOTES[NEW]: added --with_aspect_deps to blaze query, that prints additional information about aspects of target when --output is set to {xml, proto, record}.

--
MOS_MIGRATED_REVID=95272042