Default Android dependencies to a //external: label with a default binding to //tools/android: .
This is useful because we can then eventually implement an android_tools_repository() rule that lets Bazel download the Android tools from somewhere instead of requiring it to be in every workspace with Android tools.

The number of tools here is somewhat scary, therefore, I'm considering creating an android_tools rule which would have an attribute for each of these things.

Some non-trivial things about this CL:

- The labels to load are removed from AndroidConfiguration because they would resolve to e.g. //external:dx_jar, which labels just don't exist and I don't want to add dummy //external:labels not prefixed with android_
- RedirectChaser is taught how to chase redirect through bind() rules because the Android SDK is now found by //external:android_sdk -> //tools/android:sdk -> @androidsdk//:sdk . Ideally, it would be ///external:android_sdk -> @androidsdk//:sdk, but I figured I'd not fix that in this CL.

--
MOS_MIGRATED_REVID=96080553