Add all transitive module maps when we use header modules.
When we have a library 'a' that depends on a library 'm' via multiple other
libraries, and 'm' is built as a module, we need clang to load the module
map of 'm'.
There are multiple possible solutions:
1. (This CL): Add all transitive module maps as inputs when we want to use
   a module map (at any transitive level). Clang will load all module maps,
   and find the corresponding modules.
2. Add the module maps of all top-level modules (e.g. modules that are not
   reached transitively through another module) to clang's command line (we
   already pass all transtitive module maps for each library that is compiled
   as a module).

The upside of (1) is that it simplifies the module map input computation and
removes some differences between remote and local compiles (local compiles will
always see all transitive module maps).
The upside of (2) would be that we use fewer module map inputs when we do
builds that use modules as long as we only have a small subset of libraries
at the bottom of the stack compiled as modules.
Both alternatives keep transitive module maps out of the inputs for actions
that do not use header modules, thus making sure the normal case does not
regress.

We are implementing (1) because the main slow-down with transitive module maps
is during the early loading phase, and should be quickly offset by module
builds. We will revisit that decision once we have more data.

--
MOS_MIGRATED_REVID=98118092