Fix up registerEnvironmentPlistAction for bazel on Mac.
When run under bazel on my local mac, the environment variables are cleared. This causes TMPDIR to be "" which causes this script to fail. Create a tempdir and use it instead of depending on TMPDIR. Remove tempdir at end to clean up. Quote all paths just to be extra safe that somebody hasn't named their Xcode something wonky.

RELNOTES:NONE

--
MOS_MIGRATED_REVID=98640412