Refactor CcBinary / CcCommon a bit.
Don't compute the linkopts in CcCommon unconditionally, only on demand. In
order to only do this once, I had to pull out all calls to getLinkopts and
move them up in the call hierarchy. This in turn resulted in some
simplification and dead code removal in CcBinary.

--
MOS_MIGRATED_REVID=99716999