Fix race condition in SkyKey#hashCode by using the hashCode caching strategy employed by java.lang.String.
Consider the case where thread T1 calls cacheHashCode and then in the future thread T2 calls hashCode. Because the side-effects of 'cacheHashCode' were non-atomic, T2 could see a value of 'true' for 'hashCodeCached' but still see the old value (of 0) for 'hashCode' and incorrectly think the hash code is 0.

--
MOS_MIGRATED_REVID=100283097