Copies shared object dependencies when building Android native binaries.
Previously, shared objects in the transitive closure of an Android binary would be linked
in, but not be sent to the apkbuilder. This sends all the shared objects used for linking
to join the other native libraries in the APK, so they will be loadable when the main
library created by the link step is loaded.

Fixes #357.

--
MOS_MIGRATED_REVID=101365491