Rollback of commit 889f82377b69266447d0a1f8a4b37199f9f579b4.
*** Reason for rollback ***

Broke tests, I mistakenly assumed it was flakiness.

*** Original change description ***

Remove support for the deprecated include() statement.

This is part of the crusade to eliminate as every dependency in Skylark on the rest of the code so that it can be moved deeper in the dependency graph.

RELNOTES: The include() statement in BUILD files is not supported anymore.

--
MOS_MIGRATED_REVID=103284257