Adds an implicit "fully_linked" archive output to objc_library.
Adds a fully linked static library as an implicit target of objc_library. This implicit target is a library that contains the full transitive closure of libraries inside of it. For example, if liba depends on libb, then liba_fully_linked contains all of the symbols in liba and libb.

Tested:
Added unit tests that check the right libraries are linked in.

--
MOS_MIGRATED_REVID=103334108