Avoid LoadedPackageProvider in the PackageProviderForConfigs type hierarchy.
The PackageProviderForConfigurations is not a provider of loaded packages;
it loads packages unconditionally, and is executed only within SkyFunction
evaluation domains. I.e., the old type hierarchy made no sense.

Also updated the documentation and renamed getLoadedTarget to getTarget. Note
that getLoadedPackage was never called, so I removed it.

This is in preparation for removing LoadedPackageProvider.

--
MOS_MIGRATED_REVID=103754939